refactor(nfts): move platform specific top bar to headers
